Tester: Add optional include/exclude and test dir config
By default, all tests will be run for a given Tester. Add "include" and
"exclude" options to the config file "tests" section. For example:

  [tests]
  include = series/patch_count, series/cover_letter, patch/signed

If "include" is specified, only the listed tests are run. If "exclude"
is specified, all tests except the listed ones are run.

By default, the top-level "tests" directory is found in "../../", but
now this can be set via "tests" config of the "dirs" section. For example:

  [dirs]
  trees: /home/kees/code/linux-build/nipa

Signed-off-by: Kees Cook <keescook@chromium.org>